+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.14.0]
+ ### Added
+ ### Changed 
+ ### Fixed 
+- fix difference in Web template between openEhr API and etercis API  [#1470](https://github.com/ehrbase/ehrbase/pull/1470)
+
 ## [2.13.0]
  ### Added
  ### Changed 
@@ -143,3 +149,4 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 [2.11.0]: https://github.com/ehrbase/ehrbase/compare/v2.10.0...v2.11.0
 [2.12.0]: https://github.com/ehrbase/ehrbase/compare/v2.11.0...v2.12.0
 [2.13.0]: https://github.com/ehrbase/ehrbase/compare/v2.12.0...v2.13.0
+[2.14.0]: https://github.com/ehrbase/ehrbase/compare/v2.13.0...v2.14.0